uniformize uints

+++ cpp/game.cpp
@@ -12,28 +12,28 @@ using std::string;
 
 // Is there a better way to do this?
 // Arithmetic needs to be done on them
-const uint8 RL = 0;
-const uint8 RR = 1;
-const uint8 RB = 2;
-const uint8 CU = 3;
-const uint8 CD = 4;
-const uint8 CB = 5;
+const uint RL = 0;
+const uint RR = 1;
+const uint RB = 2;
+const uint CU = 3;
+const uint CD = 4;
+const uint CB = 5;
 
-const uint8 D0U = 0;
-const uint8 D0D = 1;
-const uint8 D0B = 2;
-const uint8 D1U = 3;
-const uint8 D1D = 4;
-const uint8 D1B = 5;
-const uint8 S0 = 6;
-const uint8 S1 = 7;
-const uint8 S2 = 8;
-const uint8 S3 = 9;
+const uint D0U = 0;
+const uint D0D = 1;
+const uint D0B = 2;
+const uint D1U = 3;
+const uint D1D = 4;
+const uint D1B = 5;
+const uint S0 = 6;
+const uint S1 = 7;
+const uint S2 = 8;
+const uint S3 = 9;
 
 Game::Game(): board{bitset<3*BOARD_SIZE>().set()}, frozen{bitset<BOARD_SIZE>().set()}, to_play{0},
 pieces{4, 4, 4, 4, 4, 4}, result{NONE} {}
 
-bool Game::is_empty(uint8 row, uint8 col) {
+bool Game::is_empty(uint row, uint col) {
     return !(
         board.test(row * 12 + col * 3 + 0) ||
 	board.test(row * 12 + col * 3 + 1) ||
@@ -41,7 +41,7 @@ bool Game::is_empty(uint8 row, uint8 col) {
     ); // Is there a faster way to test this given the consecutive addresses?
 }
 
-bool Game::can_place(uint8 ptype, uint8 row, uint8 col) {
+bool Game::can_place(uint ptype, uint row, uint col) {
 
     // Check if space is empty
     // This is more common than frozen spaces
@@ -71,7 +71,7 @@ bool Game::can_place(uint8 ptype, uint8 row, uint8 col) {
 
 // Returns an int representing the bottom piece of a stack, 3 if empty
 // Used to determine the legality of move moves
-int8 Game::get_bottom(uint8 row, uint8 col) {
+int8 Game::get_bottom(uint row, uint col) {
     if (board.test(row * 12 + col * 3 + 0)) return 0;
     if (board.test(row * 12 + col * 3 + 1)) return 1;
     if (board.test(row * 12 + col * 3 + 2)) return 2;
@@ -81,7 +81,7 @@ int8 Game::get_bottom(uint8 row, uint8 col) {
 
 // Returns an int representing the top of a stack, -1 if empty
 // Used to determine the legality of move moves
-int8 Game::get_top(uint8 row, uint8 col) {
+int8 Game::get_top(uint row, uint col) {
     if (board.test(row * 12 + col * 3 + 2)) return 2;
     if (board.test(row * 12 + col * 3 + 1)) return 1;
     if (board.test(row * 12 + col * 3 + 0)) return 0;
@@ -90,7 +90,7 @@ int8 Game::get_top(uint8 row, uint8 col) {
 }
 
 // Returns whether it is legal to move a stack between spaces
-bool Game::can_move(uint8 row1, uint8 col1, uint8 row2, uint8 col2) {
+bool Game::can_move(uint row1, uint col1, uint row2, uint col2) {
     // Empty spaces, move moves not possible
     if (is_empty(row1, col1) || is_empty(row2, col2)) return false;
     // Frozen spaces
@@ -100,7 +100,7 @@ bool Game::can_move(uint8 row1, uint8 col1, uint8 row2, uint8 col2) {
 }
 
 // Returns whether a move is legal
-bool Game::is_legal(uint8 move_id) {
+bool Game::is_legal(uint move_id) {
     Move move{move_id};
     // Place
     if (move.mtype) return can_place(move.ptype, move.row1, move.col1);
@@ -111,8 +111,8 @@ bool Game::is_legal(uint8 move_id) {
 // can be static (or just a normal function, although it is only used by game)
 // we could also have it not be static and change a member that tracks whether there are lines
 // since this is only called if there is a line
-void apply_line(bitset<NUM_MOVES> &legal_moves, uint8 line) {
-    legal_moves &= line_breakers[uint8];
+void apply_line(bitset<NUM_MOVES> &legal_moves, uint line) {
+    legal_moves &= line_breakers[uint];
 }
 
 // Finds lines and moves that break all lines
@@ -123,7 +123,7 @@ bool Game::get_line_breakers(bitset<NUM_MOVES> &legal_moves) {
     int8 space_0, space_1, space_2, space_3;
 
     // Rows
-    for (uint8 i = 0; i < 4; ++i) {
+    for (uint i = 0; i < 4; ++i) {
         space_1 = get_top(i, 1);
         if (space_1 != -1) { // If not empty
             space_2 = get_top(i, 2);
@@ -148,7 +148,7 @@ bool Game::get_line_breakers(bitset<NUM_MOVES> &legal_moves) {
     }
 
     // Columns
-    for (uint8 j = 0; j < 4; ++j) {
+    for (uint j = 0; j < 4; ++j) {
         space_1 = get_top(1, j);
         if (space_1 != -1) { // If not empty
             space_2 = get_top(2, j);
@@ -258,7 +258,7 @@ void Game::get_legal_moves(bitset<NUM_MOVES> &legal_moves) {
     // Filter out moves that don't break lines
     is_lines = get_line_breakers(legal_moves);
 
-    for (uint8 i = 0; i < NUM_MOVES; ++i) {
+    for (uint i = 0; i < NUM_MOVES; ++i) {
         if (legal_moves.test(i) && !is_legal(i)) {
             legal_moves.flip(i);
 	    }
@@ -285,7 +285,7 @@ void Game::get_legal_moves(bitset<NUM_MOVES> &legal_moves) {
 }
 
 // Apply move to game. Does not check for legality
-void Game::do_move(uint8 move_id) {
+void Game::do_move(uint move_id) {
     
     Move move{move_id};
 
@@ -299,7 +299,7 @@ void Game::do_move(uint8 move_id) {
     }
     // Move
     else {
-        for (uint8 i = 0; i < 3; ++i) {
+        for (uint i = 0; i < 3; ++i) {
 	        board.set(
 	            move.row2 * 12 + move.col2 * 3 + i,
 		        board.test(move.row1 * 12 + move.col1 * 3 + i) ||
@@ -316,7 +316,7 @@ bool Game::is_terminal() {
 }
 
 void Game::write_game_state(float game_state[GAME_STATE_SIZE]) {
-    for (uint8 i = 0; i < 3 * BOARD_SIZE; ++i) {
+    for (uint i = 0; i < 3 * BOARD_SIZE; ++i) {
         if (board.get(i)) {
             game_state[i] = 1.0;
         }
@@ -324,7 +324,7 @@ void Game::write_game_state(float game_state[GAME_STATE_SIZE]) {
             game_state[i] = 0.0;
         }
     }
-    for (uint8 i = 0; i < BOARD_SIZE; ++i) {
+    for (uint i = 0; i < BOARD_SIZE; ++i) {
         if (frozen.get(i)) {
             game_state[3 * BOARD_SIZE + i] = 1.0;
         }
@@ -333,7 +333,7 @@ void Game::write_game_state(float game_state[GAME_STATE_SIZE]) {
         }
     }
     // Canonize the pieces
-    for (uint8 i = 0; i < 6; ++i) {
+    for (uint i = 0; i < 6; ++i) {
         game_state[4 * BOARD_SIZE + i] = (float)pieces[(to_play * 3 + i) % 6] / 4.0;
     }
 }
