+++ cpp/node.h
@@ -14,11 +14,11 @@ using std::bitset;
     // Game state
     Game game;
     // visits is number of times this node has been searched
-    uint16 visits;
+    uint visits;
     // depth starts at 0
-    uint8 depth;
+    uint depth;
     // Index of parent node
-    uint32 parent;
+    uint parent;
     float evaluation;
     std::array<float, NUM_MOVES> probabilities;
     // legal_moves denotes which moves are legal, stores in node instead of game since Node is the one using it (?)
@@ -32,12 +32,12 @@ using std::bitset;
     Node(const Game &game);
     // Used when writing into a new node
     // Will copy a game, then apply the move
-    Node(const Game &game, uint8 depth, uint32 parent, uint8 move_choice);
+    Node(const Game &game, uint depth, uint parent, uint move_choice);
     ~Node() = default;
 
     // overwrite relevant parts of node
     // other things will be overwritten lazily
-    void overwrite(const Game &new_game, uint8 new_depth, uint32 new_parent, uint8 move_choice);
+    void overwrite(const Game &new_game, uint new_depth, uint new_parent, uint move_choice);
 
     bool is_terminal();
